finished update images form and added comments
